Unify alert engine and polling loop
