+++ app/services/alert_engine.py
@@ -1,67 +1,121 @@
-from __future__ import annotations
-import os, asyncio, datetime as dt
 from typing import Dict, Any, List, Optional
-import httpx
-from sqlalchemy.orm import Session
-from sqlalchemy import select
-from app.services.db import ensure_db
-from app.models import Alert, AlertTrigger
+import os, asyncio, time
+from app.services import alerts_store
+from app.services.providers import get_last_price
+from app.services.tradier import get_timesales_1min
 
-POLL_SEC = int(os.getenv("ALERT_POLL_SEC", "30"))
-POLY_KEY = os.getenv("POLYGON_API_KEY")
+# in-memory cache to reduce DB hits (refresh every 10s)
+_LAST_LOAD = 0.0
+_CACHE: Dict[int, Dict[str,Any]] = {}
+_LOAD_INTERVAL = 10.0
 
+# recent triggers (in-memory mirror for fast API reads)
+_RECENT: List[Dict[str,Any]] = []
+_RECENT_MAX = 100
+POLL_SEC = int(os.getenv("ALERT_POLL_SEC", "15"))
 
-async def _last_minute_close_tradier(symbol: str) -> Optional[float]:
+def _maybe_reload():
+    global _LAST_LOAD, _CACHE
+    if time.time() - _LAST_LOAD < _LOAD_INTERVAL:
+        return
+    active = alerts_store.list_active()
+    _CACHE = {a["id"]: a for a in active}
+    _LAST_LOAD = time.time()
+
+def _vwap(bars: List[Dict[str,Any]], window: int = 30) -> Optional[float]:
+    if not bars: return None
+    w = bars[-window:] if len(bars) > window else bars
+    num = 0.0; den = 0.0
+    for b in w:
+        p = (b["h"] + b["l"] + b["c"]) / 3.0
+        v = b["v"] or 0.0
+        num += p * v; den += v
+    return (num/den) if den > 0 else None
+
+def _pct_change(bars: List[Dict[str,Any]], minutes: int) -> Optional[float]:
+    if len(bars) < minutes+1: return None
+    a = bars[-minutes-1]["c"]; b = bars[-1]["c"]
+    if a == 0: return None
+    return (b - a) / a * 100.0
+
+def _last_price(bars: List[Dict[str,Any]]) -> Optional[float]:
+    return float(bars[-1]["c"]) if bars else None
+
+def _fire(alert: Dict[str,Any], symbol: str, reason: str, ctx: Dict[str,Any]):
+    alerts_store.mark_triggered(alert["id"])
+    payload = {"reason": reason, "context": ctx}
+    alerts_store.add_trigger(alert["id"], symbol, payload)
+    _RECENT.insert(0, {"alert_id": alert["id"], "symbol": symbol, "reason": reason, "payload": payload, "ts": time.time()})
+    del _RECENT[_RECENT_MAX:]
+
+def evaluate_for_symbol(symbol: str, bars: List[Dict[str,Any]]):
+    _maybe_reload()
+    if not _CACHE: return
+    lp = _last_price(bars)
+    if lp is None: return
+    vwap30 = _vwap(bars, 30)
+    for alert in list(_CACHE.values()):
+        if alert["symbol"].upper() != symbol.upper(): continue
+        cond = alert["condition"] or {}
+        t = (cond.get("type") or "").lower()
+        if t == "price_above":
+            x = float(cond.get("value", 0))
+            if lp >= x: _fire(alert, symbol, f"price_above {x}", {"last": lp})
+        elif t == "price_below":
+            x = float(cond.get("value", 0))
+            if lp <= x: _fire(alert, symbol, f"price_below {x}", {"last": lp})
+        elif t == "cross_vwap_up":
+            if vwap30 is not None and bars[-2:]:
+                prev = bars[-2]["c"]
+                if prev < vwap30 and lp >= vwap30:
+                    _fire(alert, symbol, "cross_vwap_up", {"last": lp, "vwap30": vwap30})
+        elif t == "cross_vwap_down":
+            if vwap30 is not None and bars[-2:]:
+                prev = bars[-2]["c"]
+                if prev > vwap30 and lp <= vwap30:
+                    _fire(alert, symbol, "cross_vwap_down", {"last": lp, "vwap30": vwap30})
+        elif t == "percent_change_ge":
+            mins = int(cond.get("minutes", 5))
+            th   = float(cond.get("threshold_pct", 1.0))
+            pc = _pct_change(bars, mins)
+            if pc is not None and abs(pc) >= th:
+                _fire(alert, symbol, f"percent_change_ge {th}% ({pc:.2f}%)", {"pct": pc, "minutes": mins})
+
+def recent_triggers(limit: int = 50) -> List[Dict[str,Any]]:
+    # prefer DB for durability; fall back to memory
     try:
-        from app.services.tradier import get_timesales_1min
-        js = await get_timesales_1min(symbol, minutes_back=1)
-        if js.get("ok") and js.get("bars"):
-            return float(js["bars"][-1]["c"])
+        rows = alerts_store.recent_triggers(limit)
+        if rows: return rows
     except Exception:
         pass
-    return None
+    return _RECENT[:limit]
 
-async def _last_minute_close(symbol: str) -> Optional[float]:
 
-    """Fetch the most recent 1-minute bar close for today."""
-    if not POLY_KEY: return None
-    today = dt.date.today()
-    start = today.isoformat(); end = today.isoformat()
-    url = f"https://api.polygon.io/v2/aggs/ticker/{symbol.upper()}/range/1/minute/{start}/{end}?adjusted=true&sort=desc&limit=1&apiKey={POLY_KEY}"
-    async with httpx.AsyncClient(timeout=10) as client:
-        r = await client.get(url)
-        if r.status_code != 200: return None
-        js = r.json()
-        if js.get("status") not in ("OK","DELAYED"): return None
-        res = js.get("results") or []
-        if not res: return None
-        return float(res[0].get("c"))
+async def _fetch_bars(symbol: str, minutes_back: int = 60) -> List[Dict[str,Any]]:
+    try:
+        md = await get_timesales_1min(symbol, minutes_back=minutes_back)
+        return md.get("bars") or []
+    except Exception:
+        return []
 
-def _condition_met(kind: str, value: float, price: float, cfg: Dict[str,Any]) -> bool:
-    if kind == "price_above": return price >= value
-    if kind == "price_below": return price <= value
-    return False
 
 async def poll_once():
-    engine, SessionLocal = ensure_db()
-    db: Session = SessionLocal()
-    try:
-        rows: List[Alert] = db.execute(select(Alert).where(Alert.is_active == True)).scalars().all()
-        symbols = sorted({r.symbol for r in rows})
-        prices: Dict[str, Optional[float]] = {}
-        # fetch prices sequentially to respect free-tier
-        for s in symbols:
-            prices[s] = await _last_minute_close(s)
-        for a in rows:
-            p = prices.get(a.symbol)
-            if p is None: continue
-            cond = a.condition or {}
-            k = cond.get("type"); v = cond.get("value")
-            if k and isinstance(v,(int,float)) and _condition_met(k, float(v), float(p), cond):
-                db.add(AlertTrigger(alert_id=a.id, symbol=a.symbol, payload={"price": p, "condition": a.condition}))
-        db.commit()
-    finally:
-        db.close()
+    _maybe_reload()
+    if not _CACHE:
+        return
+    symbols = sorted({a["symbol"] for a in _CACHE.values()})
+    data: Dict[str, List[Dict[str,Any]]] = {}
+    for sym in symbols:
+        bars = await _fetch_bars(sym)
+        if not bars:
+            last = await get_last_price(sym)
+            if last is not None:
+                bars = [{"c": float(last)}]
+        data[sym] = bars
+    for sym, bars in data.items():
+        if bars:
+            evaluate_for_symbol(sym, bars)
+
 
 async def run_loop():
     while True:
